perf : Performance improvements in the code by modifying the database schema, routes and add the questions.go file
