+++ backend/controllers/questions/questions-controller.go
@@ -1,16 +1,31 @@
 package controllers
 
-// func GetYearwiseQuestions(w http.ResponseWriter, r *http.Request) {
-// 	vars := mux.Vars(r)
-// 	year := vars["year"]
-// 	var question []config.Questions
-// 	result := db.Find(&question, "year = ?", year)
-// 	body, err := json.Marshal(result)
-// 	if err != nil {
-// 		fmt.Println(err)
-// 		return
-// 	}
-// 	w.Header().Set("Content-Type", "application/json")
-// 	w.WriteHeader(http.StatusOK)
-// 	w.Write(body)
-// }
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"strconv"
+
+	"github.com/gorilla/mux"
+	"prepboost.com/web/models"
+)
+
+//Get questions based on the respective year
+func GetYearwiseQuestions(w http.ResponseWriter, r *http.Request) {
+	vars := mux.Vars(r)
+	year := vars["year"]
+	YEAR, err := strconv.ParseInt(year, 0, 0)
+	questions := models.GetQuestionsByYear(YEAR)
+	if err != nil {
+		fmt.Println("error while parsing questions year")
+	}
+
+	body, err := json.Marshal(questions)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+	w.Write(body)
+}
